move FlinkDemo to example & clean log print
